+++ src/App.tsx
@@ -8,7 +8,9 @@ import Navigation from "./components/Navigation/Navigation";
 import ErrorMessage from "./components/ErrorMessage/ErrorMessage";
 import Loading from "./components/Loading/Loading";
 
-import { useAuth, useStation, useLine } from "./hooks";
+import useAuth from "./hooks/useAuth";
+import useStation from "./hooks/useStation";
+import useLine from "./hooks/useLine";
 
 import { PAGE_PATH, NAV_LINKS } from "./utils/constants/route";
