+++ src/components/ListItem/ListItem.tsx
@@ -4,7 +4,7 @@ import { FlexAlignCenter } from "../Layout/FlexContainer/FlexContainer";
 import Button from "../Button/Button";
 import Confirm from "../Confirm/Confirm";
 
-import { useModal } from "../../hooks";
+import useModal from "../../hooks/useModal";
 
 import {
   ListItemBlock,
