--- src/hooks/index.ts
@@ -1,12 +1,0 @@
-import { useAppSelector, useAppDispatch } from "./useRedux";
-
-import useAuth from "./useAuth";
-import useStation from "./useStation";
-import useLine from "./useLine";
-import useSelect from "./useSelect";
-import useModal from "./useModal";
-import useForm from "./useForm";
-
-export { useAppSelector, useAppDispatch };
-
-export { useAuth, useStation, useLine, useSelect, useModal, useForm };
