+++ src/pages/SectionManagement/SectionManagementPage.tsx
@@ -10,7 +10,10 @@ import Button from "../../components/Button/Button";
 import Select from "../../components/Select/Select";
 import ListItem from "../../components/ListItem/ListItem";
 
-import { useStation, useLine, useSelect, useModal } from "../../hooks";
+import useModal from "../../hooks/useModal";
+import useStation from "../../hooks/useStation";
+import useLine from "../../hooks/useLine";
+import useSelect from "../../hooks/useSelect";
 
 import { SIZE } from "../../utils/constants/size";
